requirements.txt
aiogram
pyyaml
python-dotenv
asyncpg
redis>=4.5.5
